Move StartInfo to Guide component folder
