+++ src/modules/PercentGame/components/Guide/StartInfo/types.ts
@@ -1,0 +1,8 @@
+import { GameIcon } from 'components/Icons/types'
+
+export interface GameStartInfo {
+  id: string
+  icon: (props: GameIcon) => JSX.Element
+  title: string
+  value: number
+}
